Modularize router structure by renaming and organizing API endpoints for devices, households, dashboard, and pets
- Remove deprecated endpoints
- Introduce enum for OpenAPI tags
