+++ surehub_api/entities/openapi.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class Tags(Enum):
+    DASHBOARD = "Dashboard"
+    HOUSEHOLD = "Household"
+    DEVICE = "Device"
+    PET = "Pet"
