+++ surehub_api/services/households.py
@@ -5,8 +5,8 @@ import requests
 from fastapi import HTTPException
 
 from surehub_api.config import settings
-from surehub_api.models import surehub
-from surehub_api.modules import auth
+from surehub_api.entities import surehub
+from surehub_api.services import auth
 
 
 def get_households() -> list:
